address some constant string warnings
svn:revision:49566
svn:branch:trunk
svn:account:n_reed

+++ src/other/step/src/clstepcore/STEPaggregate.cc
@@ -1921,7 +1921,7 @@ IntNode::STEPwrite(std::string &s, const char *)
     char tmp[BUFSIZ];
     if(value != S_INT_NULL)
     {
-	sprintf(tmp, "%d", value);
+	sprintf(tmp, "%d", (int)value);
 	s = tmp;
     }
     else
